mcp-server/requirements.txt
python-multipart==0.0.20
jsonschema-pydantic==0.6
sse-starlette==2.3.6
fastapi==0.115.12
uvicorn==0.23.1
pymongo==4.5.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
SQLAlchemy==2.0.13
loguru==0.7.2
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-milvus==0.2.0
langchain-openai==0.3.19
langchain-deepseek==0.1.3
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langmem==0.0.27
langsmith==0.3.45
fastembed==0.2.6
pdfplumber==0.11.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
openai==1.84.0